Moves timer FBs to dedicated folder
Moves the timer function blocks to a dedicated "timers"
subdirectory to improve code organization and maintainability.

This change involves updating CMakeLists.txt files to reflect the
new directory structure.

Moves blink FB to timers subfolder

Moves the blink function block definition files to a dedicated "timers" subfolder.

This change improves project organization and clarifies the location of timer-related function blocks within the codebase.

Moves E_PULSE and E_CYCLE to the timers namespace

Moves the E_PULSE and E_CYCLE event blocks to the timers namespace to better organize the event blocks.
This change aligns the directory structure with the logical grouping of timer-related event blocks.

Fixes include paths for timer events

Updates the include paths for `E_CYCLE_fbt` and `E_PULSE_fbt` to reflect their correct locations.
This resolves build issues caused by incorrect header file paths.

Refactors timer FB implementations

Refactors timer function block implementations for consistency and clarity.

Simplifies the initialization of connection parameters and improves code readability by removing redundant 'default' cases in switch statements.
