+++ modules/utils/include/forte/eclipse4diac/utils/signals/E_BLINK_TRAIN_fbt.h
@@ -30,7 +30,7 @@
 #include "forte/datatypes/forte_array.h"
 #include "forte/datatypes/forte_array_fixed.h"
 #include "forte/datatypes/forte_array_variable.h"
-#include "forte/iec61499/events/E_PULSE_fbt.h"
+#include "forte/iec61499/events/timers/E_PULSE_fbt.h"
 #include "forte/iec61499/events/E_TRAIN_fbt.h"
 #include "forte/iec61131/arithmetic/GEN_ADD_fbt.h"
 
@@ -43,7 +43,7 @@ namespace forte::eclipse4diac::utils::signals {
       static const TEventID scmEventSTOPID = 1;
       static const TEventID scmEventCNFID = 0;
 
-      forte::CInternalFB<iec61499::events::FORTE_E_PULSE> fb_E_TP;
+      forte::CInternalFB<iec61499::events::timers::FORTE_E_PULSE> fb_E_TP;
       forte::CInternalFB<iec61499::events::FORTE_E_TRAIN> fb_E_TRAIN;
       forte::CInternalFB<iec61131::arithmetic::GEN_ADD> fb_ADD_2;
