+++ stdfblib/events/include/forte/iec61499/events/timers/CMakeLists.txt
@@ -1,0 +1,23 @@
+#############################################################################
+# Copyright (c) 2025 Martin Erich Jobst
+#
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License 2.0 which is available at
+# http://www.eclipse.org/legal/epl-2.0.
+#
+# SPDX-License-Identifier: EPL-2.0
+#
+# Contributors:
+#    Martin Erich Jobst
+#      - initial API and implementation and/or initial documentation
+#############################################################################
+
+target_sources(forte-events PUBLIC
+        FILE_SET HEADERS
+        FILES
+        E_PULSE_fbt.h
+        E_TOF_fbt.h
+        E_TONOF_fbt.h
+        E_TON_fbt.h
+        E_TP_fbt.h
+)
