+++ stdfblib/events/include/forte/iec61499/events/timers/E_TONOF_fbt.h
@@ -1,5 +1,5 @@
 /*************************************************************************
- *** Copyright (c) 2023, 2024 HR Agrartechnik GmbH
+ *** Copyright (c) 2024 HR Agrartechnik GmbH
  *** This program and the accompanying materials are made available under the
  *** terms of the Eclipse Public License 2.0 which is available at
  *** http://www.eclipse.org/legal/epl-2.0.
@@ -8,39 +8,42 @@
  ***
  *** This file was generated using the 4DIAC FORTE Export Filter V1.0.x NG!
  ***
- *** Name: E_PULSE
- *** Description: standard timer function block (pulse)
+ *** Name: E_TONOF
+ *** Description: standard timer function block (on/off-delay timing)
  *** Version:
- ***     1.0: 2023-08-21/Franz Hoepfinger - HR Agrartechnik GmbH - initial implementation as E_IMPULSE
- ***     1.0: 2024-03-05/Franz Hoepfinger - HR Agrartechnik GmbH - renamed to E_PULSE
+ ***     1.0: 2024-03-04/Franz Hoepfinger - HR Agrartechnik GmbH -
  ***     1.1: 2024-04-23/Franz Hoepfinger - HR Agrartechnik GmbH - Add a Reset to Timer FBs
+ ***     3.0: 2025-04-14/Patrick Aigner -  - changed package
  *************************************************************************/
 
 #pragma once
 
 #include "forte/cfb.h"
 #include "forte/typelib.h"
-#include "forte/datatypes/forte_time.h"
 #include "forte/datatypes/forte_bool.h"
+#include "forte/datatypes/forte_time.h"
 #include "forte/iec61131_functions.h"
 #include "forte/datatypes/forte_array_common.h"
 #include "forte/datatypes/forte_array.h"
 #include "forte/datatypes/forte_array_fixed.h"
 #include "forte/datatypes/forte_array_variable.h"
+#include "forte/iec61499/events/E_SWITCH_fbt.h"
 #include "forte/iec61499/events/E_DELAY_fbt.h"
-#include "forte/iec61499/events/E_SR_fbt.h"
+#include "forte/iec61499/events/E_RS_fbt.h"
 
-namespace forte::iec61499::events {
-  class FORTE_E_PULSE final : public CCompositeFB {
-      DECLARE_FIRMWARE_FB(FORTE_E_PULSE)
+namespace forte::iec61499::events::timers {
+  class FORTE_E_TONOF final : public CCompositeFB {
+      DECLARE_FIRMWARE_FB(FORTE_E_TONOF)
 
     private:
       static const TEventID scmEventREQID = 0;
       static const TEventID scmEventRID = 1;
       static const TEventID scmEventCNFID = 0;
 
-      CInternalFB<FORTE_E_DELAY> fb_E_DELAY;
-      CInternalFB<FORTE_E_SR> fb_E_SR;
+      CInternalFB<forte::iec61499::events::FORTE_E_SWITCH> fb_E_SWITCH;
+      CInternalFB<forte::iec61499::events::FORTE_E_DELAY> fb_E_DELAY_ON;
+      CInternalFB<forte::iec61499::events::FORTE_E_RS> fb_E_RS;
+      CInternalFB<forte::iec61499::events::FORTE_E_DELAY> fb_E_DELAY_OFF;
 
       void readInputData(TEventID paEIID) override;
       void writeOutputData(TEventID paEIID) override;
@@ -48,15 +51,19 @@ namespace forte::iec61499::events {
       CDataConnection *getIf2InConUnchecked(TPortId paDIID) override;
 
     public:
-      FORTE_E_PULSE(StringId paInstanceNameId, CFBContainer &paContainer);
+      FORTE_E_TONOF(StringId paInstanceNameId, CFBContainer &paContainer);
 
       CEventConnection conn_CNF;
 
-      CDataConnection *conn_PT;
+      CDataConnection *conn_IN;
+      CDataConnection *conn_PT_ON;
+      CDataConnection *conn_PT_OFF;
 
       COutDataConnection<CIEC_BOOL> conn_Q;
 
-      COutDataConnection<CIEC_TIME> conn_if2in_PT;
+      COutDataConnection<CIEC_BOOL> conn_if2in_IN;
+      COutDataConnection<CIEC_TIME> conn_if2in_PT_ON;
+      COutDataConnection<CIEC_TIME> conn_if2in_PT_OFF;
 
       CIEC_ANY *getDI(size_t) override;
       CIEC_ANY *getDO(size_t) override;
@@ -64,4 +71,4 @@ namespace forte::iec61499::events {
       CDataConnection **getDIConUnchecked(TPortId) override;
       CDataConnection *getDOConUnchecked(TPortId) override;
   };
-} // namespace forte::iec61499::events
+} // namespace forte::iec61499::events::timers
